fixed a bug when publishing from a fresh clone

+++ publish.sh
@@ -3,6 +3,6 @@ set -e
 
 python3 -m pip install --user --upgrade setuptools wheel
 python3 -m pip install --user --upgrade twine
-rm -r ./dist/
+[ -d ./dist/ ] && rm -r ./dist/
 python3 setup.py sdist bdist_wheel
 twine upload dist/*
